+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### New Features
 - freeze: add `--reproducible` flag that zeros dynamic header metadata
 
+- Add "loop" static characteristic in Extractors
+
 ### Breaking Changes
 
 - Remove redundant `os_` parameter from `capa.loader.collect_metadata()`. External scripts calling this function must be updated to remove this argument. OS metadata now relies solely on the extractor yielding it via `extract_global_features()`. @mike-hunhoff #3076